refactor: template -> templates
