+++ database/templates.py
